# SPU LLVM: Some frest optimizations