Use vfixupimmps (avx-512 only) for spu_re_acc special cases:
Speeds up the SPU Reciprocal Accurate test from 2480ms to 2195ms

Before: 
![image](https://github.com/user-attachments/assets/ee7747ff-a2f7-4aa7-8d6c-2d89403a8448)

After:
![image](https://github.com/user-attachments/assets/e775e6dc-d6b8-46d6-824c-f99f0664e4ac)

Compute frest exponent at runtime rather than using the lookup table:
Speeds up the SPU Reciprocal Accurate test (with the spu_re_acc pattern matching disabled) from 8846ms to 8264ms
Before:
![image](https://github.com/user-attachments/assets/b82ca483-c721-4533-876f-b5741580cefc)
After:
![image](https://github.com/user-attachments/assets/8004a8a9-04c9-4502-8ee8-550d357674a7)